pyproject.toml
[build-system]
requires = ["hatchling>=1.26.0"]
build-backend = "hatchling.build"

[project]
name = "fastchain-cucc"
version = "0.1.0"
description = "Enterprise-ready FastChain distribution (CUCC)"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.12"
authors = [
  { name = "CUCC", email = "noreply@example.com" },
]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]
dependencies = [
    "aiosqlite>=0.22.1",
    "apscheduler>=3.11.2",
    "beanie>=2.0.1",
    "dynaconf>=3.2.12",
    "fastapi>=0.123.10",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "langchain-core>=1.1.2",
    "langchain-openai>=1.1.1",
    "loguru>=0.7.3",
    "motor>=3.7.1",
    "prometheus-client>=0.23.1",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "redis>=7.1.0",
    "sqlalchemy[asyncio]>=2.0.45",
    "uvicorn>=0.38.0",
]

[project.urls]
Homepage = "https://example.com"
Repository = "https://example.com"

[project.scripts]
fastchain = "fastchain.cli:main"

[tool.hatch.build]
include = [
  "fastchain/**",
  "README.md",
  "LICENSE",
  "pyproject.toml",
]
exclude = [
  "tests/**",
  "config/**",
  "log/**",
  "data/**",
  "uv.lock",
  "**/__pycache__/**",
  "**/*.pyc",
  "**/.DS_Store",
]

[tool.hatch.build.targets.wheel]
packages = ["fastchain"]
